requirements.txt
python-dotenv>=1.0.0
crawl4ai>=0.2.0
google-adk
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
requests>=2.31.0 